pyproject.toml
[tool.poetry]
name = "slavka-developer"
version = "0.1.0"
description = ""
authors = ["Marek <marek.tamaskovic@kiwi.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydicom = "^2.4.4"
click = "^8.1.7"
python-dotenv = "^1.0.1"
tqdm = "^4.66.4"

[tool.ruff]
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
